+++ src/Exceptions/SpamException.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Spatie\Honeypot\Exceptions;
+
+use Exception;
+
+class SpamException extends Exception
+{
+
+}
